+++ src/services/scriptingservice.h
@@ -50,8 +50,8 @@ public:
     Q_INVOKABLE QString downloadUrlToString(QUrl url);
     Q_INVOKABLE QString downloadUrlToMedia(QUrl url,
                                            bool returnUrlOnly = false);
-    Q_INVOKABLE QString insertMedia(QString mediaFilePath,
-                                    bool returnUrlOnly = false);
+    Q_INVOKABLE QString insertMediaFile(QString mediaFilePath,
+                                        bool returnUrlOnly = false);
     Q_INVOKABLE void registerCustomAction(
             QString identifier, QString menuText, QString buttonText = "",
             QString icon = "", bool useInNoteEditContextMenu = false,
@@ -136,6 +136,8 @@ public:
     QStringList callAutocompletionHook();
 
     Q_INVOKABLE QStringList searchTagsByName(QString name);
+
+    Q_INVOKABLE void regenerateNotePreview();
 
 private:
     QQmlEngine *_engine;